+++ packages/nwc/example/make_invoice.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
 import 'package:ndk/ndk.dart';
-import 'package:ndk_nwc/commands/make_invoice_response.dart';
 import 'package:ndk_nwc/nwc.dart';
 import 'package:ndk_nwc/nwc_connection.dart';
+import 'package:ndk_nwc/responses/make_invoice_response.dart';
 import 'package:test/test.dart';
 
 void main() {
