+++ packages/nwc/lib/nwc.dart
@@ -3,17 +3,22 @@ import 'dart:convert';
 
 import 'package:ndk/ndk.dart';
 import 'package:ndk/shared/nips/nip04/nip04.dart';
-import 'package:ndk_nwc/commands/get_balance_response.dart';
-import 'package:ndk_nwc/commands/get_info_response.dart';
-import 'package:ndk_nwc/commands/list_transactions_response.dart';
-import 'package:ndk_nwc/commands/lookup_invoice_response.dart';
-import 'package:ndk_nwc/commands/pay_invoice_response.dart';
-import 'package:ndk_nwc/nosrt_wallet_connect_uri.dart';
-import 'package:ndk_nwc/nwc_connection.dart';
-import 'package:ndk_nwc/requests/nwc_request.dart';
-
-import 'commands/make_invoice_response.dart';
-import 'commands/nwc_response.dart';
+import 'responses/get_balance_response.dart';
+import 'responses/get_info_response.dart';
+import 'responses/list_transactions_response.dart';
+import 'responses/lookup_invoice_response.dart';
+import 'responses/pay_invoice_response.dart';
+import 'nosrt_wallet_connect_uri.dart';
+import 'nwc_connection.dart';
+import 'requests/get_balance.dart';
+import 'requests/get_info.dart';
+import 'requests/list_transactions.dart';
+import 'requests/lookup_invoice.dart';
+import 'requests/make_invoice.dart';
+import 'requests/nwc_request.dart';
+import 'requests/pay_invoice.dart';
+import 'responses/make_invoice_response.dart';
+import 'responses/nwc_response.dart';
 import 'consts/nwc_kind.dart';
 import 'consts/transaction_type.dart';
 import 'consts/nwc_method.dart';
