+++ packages/nwc/lib/nwc_connection.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 
 import 'package:bip340/bip340.dart';
 import 'package:ndk/ndk.dart';
-import 'package:ndk_nwc/commands/get_info_response.dart';
-import 'package:ndk_nwc/commands/nwc_response.dart';
+import 'responses/get_info_response.dart';
+import 'responses/nwc_response.dart';
 import 'package:ndk_nwc/nosrt_wallet_connect_uri.dart';
 
 class NwcConnection {
