+++ packages/nwc/lib/requests/get_balance.dart
@@ -1,0 +1,8 @@
+import 'package:ndk_nwc/consts/nwc_method.dart';
+
+import 'nwc_request.dart';
+
+// Subclass for requests to get balance
+class GetBalanceRequest extends NwcRequest {
+  const GetBalanceRequest() : super(method: NwcMethod.GET_BALANCE);
+}
