+++ packages/nwc/lib/requests/get_info.dart
@@ -1,0 +1,8 @@
+import 'package:ndk_nwc/consts/nwc_method.dart';
+
+import 'nwc_request.dart';
+
+// Subclass for requests to get info like supported methods
+class GetInfoRequest extends NwcRequest {
+  const GetInfoRequest() : super(method: NwcMethod.GET_INFO);
+}
