+++ packages/nwc/lib/responses/get_info_response.dart
