+++ packages/nwc/lib/responses/lookup_invoice_response.dart
@@ -1,9 +1,9 @@
 // ignore_for_file: camel_case_types
 
-import 'package:ndk_nwc/commands/nwc_response.dart';
+import 'nwc_response.dart';
 
-/// Represents the result of a 'make_invoice' response.
-class MakeInvoiceResponse extends NwcResponse {
+/// Represents the result of a 'lookup_invoice' response.
+class LookupInvoiceResponse extends NwcResponse {
   /// The type of the invoice "incoming" for invoices, "outgoing" for payments.
   final String type;
 
@@ -23,9 +23,7 @@ class MakeInvoiceResponse extends NwcResponse {
   final String paymentHash;
 
   /// The amount of the invoice (in MSATs)
-  final int amountMsat;
-
-  get amountSat => amountMsat ~/ 1000;
+  final int amount;
 
   /// The fees paid for the invoice (in MSATs).
   final int feesPaid;
@@ -37,17 +35,16 @@ class MakeInvoiceResponse extends NwcResponse {
   final int expiresAt;
 
   /// The timestamp when the invoice was settled (optional).
-
   final int? settledAt;
 
-  MakeInvoiceResponse({
+  LookupInvoiceResponse({
     required this.type,
     required this.invoice,
     required this.description,
     required this.descriptionHash,
     required this.preimage,
     required this.paymentHash,
-    required this.amountMsat,
+    required this.amount,
     required this.feesPaid,
     required this.createdAt,
     required this.expiresAt,
@@ -55,21 +52,21 @@ class MakeInvoiceResponse extends NwcResponse {
     required super.resultType,
   });
 
-  factory MakeInvoiceResponse.deserialize(Map<String, dynamic> input) {
+  factory LookupInvoiceResponse.deserialize(Map<String, dynamic> input) {
     if (!input.containsKey('result')) {
       throw Exception('Invalid input');
     }
 
     Map<String, dynamic> result = input['result'] as Map<String, dynamic>;
 
-    return MakeInvoiceResponse(
+    return LookupInvoiceResponse(
       type: result['type'] as String,
       invoice: result['invoice'] as String,
       description: result['description'] as String,
       descriptionHash: result['description_hash'] as String,
       preimage: result['preimage'] as String,
       paymentHash: result['payment_hash'] as String,
-      amountMsat: result['amount'] as int,
+      amount: result['amount'] as int,
       feesPaid: result['fees_paid'] as int,
       createdAt: result['created_at'] as int,
       expiresAt: result['expires_at'] as int,
