+++ packages/nwc/lib/responses/make_invoice_response.dart
@@ -1,9 +1,9 @@
 // ignore_for_file: camel_case_types
 
-import 'package:ndk_nwc/commands/nwc_response.dart';
+import 'nwc_response.dart';
 
-/// Represents the result of a 'lookup_invoice' response.
-class LookupInvoiceResponse extends NwcResponse {
+/// Represents the result of a 'make_invoice' response.
+class MakeInvoiceResponse extends NwcResponse {
   /// The type of the invoice "incoming" for invoices, "outgoing" for payments.
   final String type;
 
@@ -23,7 +23,9 @@ class LookupInvoiceResponse extends NwcResponse {
   final String paymentHash;
 
   /// The amount of the invoice (in MSATs)
-  final int amount;
+  final int amountMsat;
+
+  get amountSat => amountMsat ~/ 1000;
 
   /// The fees paid for the invoice (in MSATs).
   final int feesPaid;
@@ -35,16 +37,17 @@ class LookupInvoiceResponse extends NwcResponse {
   final int expiresAt;
 
   /// The timestamp when the invoice was settled (optional).
+
   final int? settledAt;
 
-  LookupInvoiceResponse({
+  MakeInvoiceResponse({
     required this.type,
     required this.invoice,
     required this.description,
     required this.descriptionHash,
     required this.preimage,
     required this.paymentHash,
-    required this.amount,
+    required this.amountMsat,
     required this.feesPaid,
     required this.createdAt,
     required this.expiresAt,
@@ -52,21 +55,21 @@ class LookupInvoiceResponse extends NwcResponse {
     required super.resultType,
   });
 
-  factory LookupInvoiceResponse.deserialize(Map<String, dynamic> input) {
+  factory MakeInvoiceResponse.deserialize(Map<String, dynamic> input) {
     if (!input.containsKey('result')) {
       throw Exception('Invalid input');
     }
 
     Map<String, dynamic> result = input['result'] as Map<String, dynamic>;
 
-    return LookupInvoiceResponse(
+    return MakeInvoiceResponse(
       type: result['type'] as String,
       invoice: result['invoice'] as String,
       description: result['description'] as String,
       descriptionHash: result['description_hash'] as String,
       preimage: result['preimage'] as String,
       paymentHash: result['payment_hash'] as String,
-      amount: result['amount'] as int,
+      amountMsat: result['amount'] as int,
       feesPaid: result['fees_paid'] as int,
       createdAt: result['created_at'] as int,
       expiresAt: result['expires_at'] as int,
