+++ packages/nwc/lib/responses/nwc_response.dart
