+++ packages/nwc/lib/responses/pay_invoice_response.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: camel_case_types
 
-import 'package:ndk_nwc/commands/nwc_response.dart';
+import 'nwc_response.dart';
 
 /// Represents the result of an pay_invoice response.
 class PayInvoiceResponse extends NwcResponse {
